zpp-processor/requirements.txt
Flask==3.1.0
gunicorn==21.2.0
pymongo==4.13.2
pandas==2.3.1
openpyxl==3.1.2
python-dotenv==1.1.1
numpy==2.3.2
